+++ inst/trelliscopeViewer/server/cogState.R
@@ -13,57 +13,115 @@
 # we do this with a function updateControlsExposedState()
 # this calls methods for each individual control
 
+filterState <- reactive({
+   fs <- input$filterStateInput
+   if(is.null(fs)) {
+      fs <- currentDisplay()$cdo$state$filter
+   } else {
+      class(fs) <- c(class(fs), "filterState")
+   }
+   logMsg("- filter state changed: ", toHash(fs))
+   fs
+})
+
+sortState <- reactive({
+   ss <- input$sortStateInput
+   if(is.null(ss)) {
+      ss <- currentDisplay()$cdo$state$sort
+   } else {
+      class(ss) <- c(class(ss), "sortState")
+   }
+   logMsg("- panel sort state changed: ", toHash(ss))
+   ss
+})
+
+panelLabelState <- reactive({
+   pl <- input$panelLabelStateInput
+   if(is.null(pl)) {
+      pl <- currentDisplay()$cdo$state$labels
+   } else {
+      class(pl) <- c(class(pl), "labelsState")
+      if(pl[1] == "__none__")
+         pl <- NULL
+   }
+   logMsg("- panel label state changed: ", toHash(pl))
+   pl
+})
+
+panelLayoutState <- reactive({
+   pls <- input$panelLayoutStateInput
+   if(is.null(pls)) {
+      pls <- currentDisplay()$cdo$state$layout      
+   } else {
+      class(pls) <- c(class(pls), "layoutState")
+   }
+   logMsg("- panel layout state changed: nrow: ", pls$nrow, ", ncol: ", pls$ncol, ", w: ", pls$w, " h: ", pls$h, " arrange: ", pls$arrange)
+   pls
+})
+
+activeCogState <- reactive({
+   ac <- input$activeCogStateInput
+   if(!is.null(ac)) {
+      class(ac) <- c(class(ac), "activeCogState")
+   }
+   logMsg("- active cog state changed: ", paste(ac, collapse=","))
+   ac
+})
+
+relatedDisplayState <- reactive({
+   rd <- input$relatedDisplayStateInput
+   if(length(rd) > 0) {
+      logMsg("- related display state changed")      
+   } else {
+      rd <- NULL
+   }
+   rd
+})
+
 cdoExposedCogState <- reactive({
    cdo <- currentDisplay()$cdo
-   filterState <- input$filterStateInput
-   if(!is.null(filterState)) {
-      logMsg("- filter state changed")
-      cdo$state$filter <- filterState      
-   }
    
-   sortState <- input$sortStateInput
-   if(!is.null(sortState)) {
-      logMsg("- panel sort state changed")
-      cdo$state$sort <- sortState      
-   }
-   
-   panelLabelState <- input$panelLabelStateInput
-   if(!is.null(panelLabelState)) {
-      logMsg("- panel label state changed: ", paste(panelLabelState, collapse=","))
-      if(any(panelLabelState == "__none__"))
-         panelLabelState <- NULL
-      cdo$state$panelLabel <- panelLabelState      
-   }
-   
-   pls <- input$panelLayoutStateInput
-   if(!is.null(pls)) {
-      logMsg("- panel layout state changed: nrow: ", pls$nrow, ", ncol: ", pls$ncol, ", w: ", pls$w, " h: ", pls$h, " arrange: ", pls$arrange)
-      cdo$state$panelLayout <- pls
-   }
-   
-   activeCogState <- input$activeCogStateInput
-   if(!is.null(activeCogState)) {
-      logMsg("- active cog state changed: ", paste(activeCogState, collapse=","))
-      cdo$state$activeCog <- activeCogState
-   }
-   
-   relatedDisplayState <- input$relatedDisplayStateInput
-   if(length(relatedDisplayState) > 0) {
-      # load the additional displays
-      logMsg("- related display state changed")
+   if(!is.null(cdo)) {
+      cdo$state$filter <- filterState()
+      cdo$state$sort <- sortState()
+      cdo$state$labels <- panelLabelState()
+      cdo$state$layout <- panelLayoutState()
+      cdo$state$activeCog <- activeCogState()
+      cdo$state$relatedDisplays <- relatedDisplayState()
+      
       relatedDisplayObjects <- list()
-      for(i in seq_along(relatedDisplayState)) {
-         curName <- relatedDisplayState[[i]]$name
-         curGroup <- relatedDisplayState[[i]]$group
-         dispKey <- paste(curGroup, curName, sep = "___")
-         if(curName == cdo$name && curGroup == cdo$group) {
-            relatedDisplayObjects[[dispKey]] <- NULL
-         } else {
-            relatedDisplayObjects[[dispKey]] <- getDisplay(name = curName, group = curGroup)
+      if(length(cdo$state$relatedDisplays) > 0) {
+         # load the additional displays
+         rd <- cdo$state$relatedDisplays
+         for(i in seq_along(rd)) {
+            curName <- rd[[i]]$name
+            curGroup <- rd[[i]]$group
+            dispKey <- paste(curGroup, curName, sep = "___")
+            if(curName == cdo$name && curGroup == cdo$group) {
+               relatedDisplayObjects[[dispKey]] <- NULL
+            } else {
+               tmp <- getDisplay(name = curName, group = curGroup)
+               
+               # load packages and data
+               if(!is.null(tmp$relatedPackages)) {
+                  logMsg("Loading packages: ", paste(tmp$relatedPackages, collapse = ", "))
+                  for(pkg in tmp$relatedPackages)
+                     suppressMessages(require(pkg, character.only = TRUE))
+               }
+               
+               # load any related data into global environment
+               # note: these should not be put in global environment
+               # but a display-specific environment - need to update
+               if(!is.null(tmp$relatedData)) {
+                  for(nm in names(tmp$relatedData))
+                     .GlobalEnv[[nm]] <- tmp$relatedData[[nm]]
+               }
+               
+               relatedDisplayObjects[[dispKey]] <- tmp
+            }
          }
       }
-      cdo$relatedDisplayObjects <- relatedDisplayObjects
-      cdo$state$relatedDisplays <- relatedDisplayState
+      cdo$relatedDisplayObjects <- relatedDisplayObjects      
    }
    
    # cdo$state$sample <- 
@@ -72,13 +130,11 @@ cdoExposedCogState <- reactive({
    cdo
 })
 
-# is this really necessary? could just set it in the browser
-# before triggering each state input
 output$exposedStateDataOutput <- renderData({
    cdo <- cdoExposedCogState()
    # if(!is.null(cdo$state))
    #    browser()
-   cdo$state
+   c(list(name = cdo$name, group = cdo$group), cdo$state)
 })
 
 output$cogBreadcrumbOutput <- renderDataLite({
@@ -93,10 +149,23 @@ output$cogBreadcrumbOutput <- renderDataLite({
          list(name = nm, class = "btn-success filter-breadcrumb", icon = "icon-filter", type = "filter")
       })
       
+      iconLookup <- list(
+         "asc" = list(
+            "numeric" = "icon-sort-numeric-asc",
+            "factor" = "icon-sort-alpha-asc"
+         ),
+         "desc" = list(
+            "numeric" = "icon-sort-numeric-desc",
+            "factor" = "icon-sort-alpha-desc"
+         )
+      )
+      
       sortNm <- names(state$sort)
       sorts <- lapply(sortNm, function(nm) {
          cur <- state$sort[[nm]]
-         list(name = nm, class = "btn-primary sort-breadcrumb", icon = cur$bcIcon, order = cur$order, type = "sort")
+         type <- cdo$cogInfo$type[cdo$cogInfo$name == nm]
+         bcIcon <- iconLookup[[cur$dir]][[type]]
+         list(name = nm, class = "btn-primary sort-breadcrumb", icon = bcIcon, order = cur$order, type = "sort")
       })
       
       if(length(sorts) > 0) {
@@ -107,6 +176,15 @@ output$cogBreadcrumbOutput <- renderDataLite({
       return(list(buttons = c(filters, sorts)))
    }
 })
+
+output$appHashOutput <- renderText({
+   cdo <- cdoExposedCogState()
+   
+   if(!is.null(cdo))
+      makeStateHash(cdo$state, cdo$name, cdo$group)
+})
+
+