+++ inst/trelliscopeViewer/server/currentDisplay.R
@@ -7,11 +7,49 @@ output$displayListOutput <- renderDataLite({
 })
 outputOptions(output, "displayListOutput", suspendWhenHidden = FALSE)
 
+appHash <- reactive({
+   str <- input$appHashInput
+   if(!is.null(str)) {
+      if(substr(str, 1, 1) == "#")
+         str <- substr(str, 2, nchar(str))
+      res <- parseQueryString(str)
+      cls <- paste(names(res), "Hash", sep = "")
+      for(i in seq_along(res))
+         class(res[[i]]) <- c("character", cls[i])
+      return(res)
+   }
+})
+
 ## holds name and group of currently selected display
 selectedDisplay <- reactive({
-   # TODO: appHash stuff here...
-   input$displaySelectInput
+   # priority: input, appHash, options
+   sld <- input$displaySelectInput
+   
+   if(is.null(sld)) {
+      sld <- appHash()
+      if(!is.null(sld)) {
+         sld$labels <- fromHash(sld$labels)
+         sld$layout <- fromHash(sld$layout)
+         sld$sort <- fromHash(sld$sort)
+         sld$filter <- fromHash(sld$filter)
+      }
+   }
+   
+   if(is.null(sld)) {
+      sld <- getOption("trsCurrentViewState")
+   }
+   
+   sld
 })
+
+# http://127.0.0.1:8100/#name=list_sold_vs_time&group=common&layout=ncol:2&sort=state:asc,slope:desc&filter=county:(regex:a),state:(select:AL;AR),slope:(from:0;to:1),meanList:(from:50)&labels=county,state,slope
+
+
+# if(!is.null(currentViewState$name))
+#    sld <- list(name = currentViewState$name,
+#       group = currentViewState$group)
+# options(trsCurrentViewState = list(group = "common", name = "list_sold_vs_time"))
+
 
 ## name of the display printed in the header
 output$headerDisplayNameOutput <- renderText({
@@ -26,7 +64,7 @@ currentDisplay <- reactive({
    
    if(!is.null(sld)) {
       logMsg("Loading display: ", paste(sld$group, "/", sld$name))
-      cdo <- do.call(getDisplay, sld)
+      cdo <- do.call(getDisplay, sld[c("name", "group")])
       logMsg("Display loaded")
       
       # load required packages
@@ -37,8 +75,8 @@ currentDisplay <- reactive({
       }
       
       # set up an environment for this display to be evaluated in
-      dispEnv <- new.env(parent = .GlobalEnv)
-      cdo$envir <- .GlobalEnv
+      # dispEnv <- new.env(parent = .GlobalEnv)
+      # cdo$envir <- .GlobalEnv
       
       # load any related data into global environment
       # note: these should not be put in global environment
@@ -52,14 +90,26 @@ currentDisplay <- reactive({
          cdo$cogDistns <- trelliscope:::getCogDistns(cdo$cogDatConn)
       
       logMsg("Getting default state...")
-      # default state values if not specified
-      if(is.null(cdo$state$panelLabelState)) {
+      
+      cdo$state$labels <- sld$labels
+      if(is.null(cdo$state$labels)) {
          defaultLabels <- cdo$cogInfo$name[cdo$cogInfo$defLabel]
+         class(defaultLabels) <- c(class(defaultLabels), "labelsState")
          if(length(defaultLabels) == 0)
             defaultLabels <- NULL
-         cdo$state$panelLabel <- defaultLabels
+         cdo$state$labels <- defaultLabels
       }
-
+      
+      cdo$state$layout <- sld$layout
+      if(is.null(cdo$state$layout)) {
+         lyt <- list(nrow = 1, ncol = 1, arrange = "row")
+         class(lyt) <- c("list", "layoutState")
+         cdo$state$layout <- lyt
+      }
+      
+      cdo$state$sort <- sld$sort
+      cdo$state$filter <- sld$filter
+      
       if(is.null(cdo$state$activeCog)) {
          defaultActive <- cdo$cogInfo$name[cdo$cogInfo$defActive]
          if(length(defaultActive) == 0)
@@ -67,10 +117,9 @@ currentDisplay <- reactive({
          cdo$state$activeCog <- defaultActive
       }
       
-      if(is.null(cdo$state$panelLayout))
-         cdo$state$panelLayout <- list(nrow = 1, ncol = 1)
+      # if(!is.null(currentViewState))
+      #    options(trsCurrentViewState = NULL)
       
-      # TODO: add related
       list(cdo = cdo)
    }
 })
@@ -125,7 +174,3 @@ output$cogMapOutput <- renderDataLite({
    cogBiFilterControlsOutputData(currentDisplay())
 })
 
-cdoCogState <- reactive({
-   cdo <- currentDisplay()$cdo
-   cdo
-})