+++ inst/trelliscopeViewer/www/assets/custom/controls-panel-labels.js
@@ -53,12 +53,12 @@ function panelLabelListSetFromExposedState() {
    });
    
    if(state) {
-      if(state.panelLabel) {
+      if(state.labels) {
          // console.log(state);
-         if(!(state.panelLabel instanceof Array)) {
-            state.panelLabel = [state.panelLabel];
+         if(!(state.labels instanceof Array)) {
+            state.labels = [state.labels];
          }
-         $.each(state.panelLabel, function(key, value) {
+         $.each(state.labels, function(key, value) {
             $("#panel-labels-select-" + value).addClass("active");
          });
       }      
@@ -66,6 +66,10 @@ function panelLabelListSetFromExposedState() {
 }
 
 function panelLabelListOutputPostRender() {
+   // console.log($(".panel-labels-select.active").length);
+   
+   $("#openModal").modal("hide");
+   
    $.getScript("assets/custom/selectables-panel-labels.js");
    
    // $("#panel-rows").trigger("change");